fixes graphql configs, keeps performance mode setting in context.config, adds the ability to update config through graphql, and refactors graphql types to move common dependencies away from local-only typeDefs
