+++ packages/graphql/src/configs/localhost.js
@@ -14,10 +14,11 @@ const config = {
   ipfsRPC: `http://${HOST}:5002`,
   relayer: `http://${HOST}:5100`,
   bridge: 'https://bridge.dev.originprotocol.com',
+  discovery: `http://${HOST}:4000/graphql`,
   notifications: `http://${HOST}:3456`,
   //growth: 'http://localhost:4001',
   performanceMode: false,
-  graphql: 'http://${HOST}:4002',
+  graphql: `http://${HOST}:4002`,
   automine: 2000,
   attestationIssuer: '0x5be37555816d258f5e316e0f84D59335DB2400B2',
   affiliate: addresses.Affiliate,
