+++ packages/graphql/src/configs/localsecure.js
@@ -13,10 +13,11 @@ const config = {
   ipfsGateway: `https://${HOST}:8081`,
   ipfsRPC: `https://${HOST}:5003`,
   bridge: 'https://bridge.dev.originprotocol.com',
+  discovery: `https://${HOST}:4000/graphql`,
   notifications: `http://${HOST}:3456`,
   //growth: 'http://localhost:4001',
   performanceMode: false,
-  graphql: 'https://${HOST}:4002',
+  graphql: `https://${HOST}:4002`,
   automine: 2000,
   attestationIssuer: '0x02c14Ee67799e1dFac9f776F2f4D5dC69Ab3Ee23',
   affiliate: addresses.Affiliate,
