+++ packages/graphql/src/configs/test.js
@@ -14,7 +14,7 @@ const config = {
   ipfsRPC: `http://${HOST}:5002`,
   bridge: 'https://bridge.dev.originprotocol.com',
   performanceMode: false,
-  graphql: 'http://${HOST}:4002',
+  graphql: `http://${HOST}:4002`,
   automine: 500,
   attestationIssuer: '0x5be37555816d258f5e316e0f84D59335DB2400B2',
