+++ packages/graphql/src/configs/truffle.js
@@ -15,7 +15,7 @@ const config = {
   growth: `http://${HOST}:4001`,
   bridge: `http://${HOST}:5000`,
   performanceMode: false,
-  graphql: 'https://${HOST}:4002',
+  graphql: `http://${HOST}:4002`,
   automine: 2000,
   attestationIssuer: '0x99C03fBb0C995ff1160133A8bd210D0E77bCD101',
   messaging: {
