+++ packages/graphql/src/typeDefs/Web3.js
@@ -1,5 +1,5 @@
 module.exports = `
-  type Subscription {
+  extend type Subscription {
     newBlock: Block
     newTransaction: NewTransaction
     transactionUpdated: TransactionUpdate
@@ -18,7 +18,7 @@ module.exports = `
     node: Transaction!
   }
 
-  type Query {
+  extend type Query {
     web3: Web3
     config: String
     contracts: [Contract]
@@ -36,6 +36,7 @@ module.exports = `
     ipfsGateway: String
     provider: String
     providerWS: String
+    performanceMode: Boolean
   }
 
   type Config {
@@ -52,9 +53,10 @@ module.exports = `
     provider: String
     providerWS: String
     relayer: String
+    performanceMode: Boolean
   }
 
-  type Mutation {
+  extend type Mutation {
     refetch: Boolean
     setNetwork(network: String, customConfig: ConfigInput): Boolean
     toggleMetaMask(enabled: Boolean): Boolean
@@ -97,32 +99,6 @@ module.exports = `
     primaryAccount: Account
   }
 
-  type Account {
-    id: ID!
-    checksumAddress: String
-    balance: Balance
-    role: String
-    name: String
-    token(symbol: String!): TokenHolder
-    identity: Identity
-    owner: Account
-    proxy: Account
-  }
-
-  type Balance {
-    wei: String
-    eth: String
-    usd: String
-  }
-
-  type TokenHolder {
-    id: ID!
-    account: Account
-    token: Token
-    balance: String
-    allowance(contract: String): String
-  }
-
   type Contract {
     id: ID!
     balance: Balance
@@ -139,114 +115,6 @@ module.exports = `
   type TransferTokenOutput {
     to: TokenHolder
     from: TokenHolder
-  }
-
-  type Transaction {
-    id: ID!
-    status: String
-    blockHash: String
-    blockNumber: Int
-    from: String
-    gas: Int
-    gasPrice: String
-    hash: String
-    input: String
-    nonce: Int
-    to: String
-    value: String
-    pct: Float
-    receipt: TransactionReceipt
-
-    # Timestamp transaction originally submitted
-    submittedAt: Int
-  }
-
-  type TransactionReceipt {
-    id: ID!
-    blockHash: String
-    blockNumber: Int
-    contractAddress: String
-    cumulativeGasUsed: Int
-    gasUsed: Int
-    logs: [Log]
-    events: [Event]
-    logsBloom: String
-    status: Boolean
-    transactionHash: String
-    transactionIndex: Int
-  }
-
-  type Log {
-    id: ID!
-    address: String
-    blockHash: String
-    blockNumber: Int
-    data: String
-    logIndex: Int
-    topics: [String]
-    transactionHash: String
-    transactionIndex: Int
-    type: String
-  }
-
-  type Event {
-    id: ID!
-    address: String
-    blockHash: String
-    blockNumber: Int
-    block: Block
-    event: String
-    logIndex: Int
-    raw: EventRaw
-    returnValues: EventReturnValues
-    returnValuesArr: [EventReturnValuesArr]
-    signature: String
-    transactionHash: String
-    transactionIndex: Int
-    type: String
-    timestamp: Int
-  }
-
-  type EventRaw {
-    data: String
-    topics: [String]
-  }
-
-  type EventReturnValues {
-    listingID: ID
-    offerID: ID
-    party: String!
-    ipfsHash: String!
-    ipfsUrl: String
-  }
-
-  type EventReturnValuesArr {
-    field: String
-    value: String
-  }
-
-  type Block {
-    id: ID!
-    number: Int
-    hash: String
-    parentHash: String
-    mixHash: String
-    nonce: String
-    sha3Uncles: String
-    logsBloom: String
-    transactionsRoot: String
-    stateRoot: String
-    receiptsRoot: String
-    miner: String
-    difficulty: String
-    totalDifficulty: String
-    extraData: String
-    size: Int
-    gasLimit: Int
-    gasUsed: Int
-    timestamp: Int
-    transactions: [String]
-    uncles: [String]
   }
 
   input WalletInput {
