+++ packages/graphql/src/typeDefs/remote.js
@@ -1,21 +1,17 @@
 import Common from './Common'
-import Web3 from './Web3'
+import Web3Common from './Web3Common'
 import Marketplace from './Marketplace'
 import Identity from './Identity'
 import Attestations from './Attestations'
 import Notifications from './Notifications'
-import CreatorConfig from './CreatorConfig'
 import Currency from './Currency'
-import Uniswap from './Uniswap'
 
 export default [
-  Common,
-  Web3,
-  Marketplace,
+  Web3Common,
   Identity,
+  Common,
+  Marketplace,
   Notifications,
   Attestations,
-  CreatorConfig,
   Currency,
-  Uniswap
 ]
